Extract connector dialog configuration
